Move tour demo to listing page and trigger it on a button click.
