+++ static/js/publisher/tour.js
@@ -3,10 +3,10 @@ import ReactDOM from "react-dom";
 
 import Tour from "./tour/tour";
 
-export default function initTour() {
-  const tour = document.createElement("div");
-  document.body.appendChild(tour);
-
-  const target = document.querySelector("h2");
-  ReactDOM.render(<Tour target={target} />, tour);
+export default function initTour({ container, target }) {
+  if (document.contains(container) && document.contains(target)) {
+    ReactDOM.render(<Tour target={target} />, container);
+  } else {
+    throw Error("initTour container or target elements not found in document");
+  }
 }
