+++ static/sass/_utilities_flex.scss
@@ -4,4 +4,9 @@
   .u-flex {
     display: flex;
   }
+
+  .u-flex--space-between {
+    display: flex;
+    justify-content: space-between;
+  }
 }
