+++ templates/index.html
@@ -226,7 +226,5 @@
       snapcraft.public.nps();
     {% endif %}
   });
-
-  snapcraft.publisher.initTour();
 </script>
 {% endblock %}
