+++ templates/publisher/_header.html
@@ -16,6 +16,7 @@
       <ul class="p-tabs__list u-float-right u-no-margin--bottom" role="tablist">
         <li class="p-tabs__item" role="presentation">
           <a
+            data-tour="listing-intro"
             href="/{{ snap_name }}/listing"
             class="p-tabs__link"
             tabindex="0"
